requirements.txt
django
pyyaml
django-grappelli